+++ pricing/src/epilot/models/shared/error.py
@@ -11,7 +11,6 @@ from typing import Optional
 
 @dataclasses.dataclass
 class Error:
-    r"""Invalid payload"""
     message: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('message') }})
     r"""Error message"""
     cause: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('cause'), 'exclude': lambda f: f is None }})